Initial commit of iTunes plugin client.
I admit, when I started looking over the original code for this, I
couldn't bear to commit it as-is. I ended up re-writing the hotpatching
code, and making various other fixes.

Right now, unfortunately, the album art transmission is mostly broken. It
appears that iTunes made a plugin api breaking change. I'll have to fix
that...